# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VEICULAR",
        "description": "Api para ERP de manutenção veicular",
        "version": "1.0.0"
    },
    "paths": {
        "/pessoas": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Lista todas as pessoas",
                "operationId": "3dc616a61eacfedcb301f11adb7fcaa1",
                "responses": {
                    "200": {
                        "description": "Lista de todas as pessoas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pessoa"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao obter a lista de pessoas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Cria uma nova pessoa",
                "operationId": "0269fc659827ad5bbdcde3393b59309b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Pessoa"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pessoa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pessoa"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro ao criar uma nova pessoa"
                    }
                }
            }
        },
        "/pessoas/{id}": {
            "get": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Lista uma pessoa específica",
                "operationId": "ed509b9d131a64e29eb74abcffb6f579",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pessoa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pessoa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pessoa não encontrada"
                    },
                    "500": {
                        "description": "Erro ao obter a pessoa"
                    }
                }
            }
        },
        "/pessoas/{pessoa}": {
            "put": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Atualiza uma pessoa",
                "operationId": "340e0612deef52b5517ef667e5dcf6ef",
                "parameters": [
                    {
                        "name": "pessoa",
                        "in": "path",
                        "description": "Pessoa a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Pessoa"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pessoa atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pessoa"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro ao atualizar a pessoa"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pessoas"
                ],
                "summary": "Deleta uma pessoa",
                "operationId": "d0f995eb8bf727d7fd204f1d877f624d",
                "parameters": [
                    {
                        "name": "pessoa",
                        "in": "path",
                        "description": "Pessoa a ser deletada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pessoa deletada com sucesso"
                    },
                    "500": {
                        "description": "Erro ao deletar a pessoa"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Pessoa": {
                "title": "Pessoa",
                "description": "Pessoa model",
                "properties": {
                    "": {
                        "title": "Veículos",
                        "description": "Relacionamento com os veículos da pessoa",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Veiculo"
                        }
                    },
                    "fillable": {
                        "title": "ID",
                        "description": "ID da pessoa",
                        "format": "int64"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Pessoa"
                }
            },
            "RevisaoVeicular": {
                "title": "RevisaoVeicular",
                "description": "RevisaoVeicular model",
                "properties": {
                    "veiculo": {
                        "$ref": "#/components/schemas/Veiculo"
                    },
                    "pessoa": {
                        "$ref": "#/components/schemas/Pessoa"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RevisaoVeicular"
                }
            },
            "Veiculo": {
                "title": "Veiculo",
                "description": "Objeto de Veiculo",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "pessoa_id": {
                        "type": "integer"
                    },
                    "marca": {
                        "type": "string"
                    },
                    "tipo": {
                        "type": "string"
                    },
                    "modelo": {
                        "type": "string"
                    },
                    "placa": {
                        "type": "string"
                    },
                    "pessoa": {
                        "$ref": "#/components/schemas/Pessoa"
                    },
                    "revisoes": {
                        "$ref": "#/components/schemas/RevisaoVeicular"
                    },
                    "fillable": {
                        "title": "ID",
                        "description": "ID do veículo",
                        "format": "int64"
                    }
                },
                "type": "object"
            }
        }
    },
...
}